+++ pixelbin/platform/models/UpdateFileRequest.py
@@ -26,7 +26,7 @@ class UpdateFileRequest(BaseSchema):
     
     path = fields.Str(required=False)
     
-    access = fields.Str(required=False)
+    access = fields.Str(required=False, validate=OneOf([val.value for val in AccessEnum.__members__.values()]))
     
     isActive = fields.Boolean(required=False)